+++ src/project/client/editor/SubmitServiceAsync.java
@@ -4,5 +4,5 @@ import com.google.gwt.user.client.rpc.AsyncCallback;
 
 public interface SubmitServiceAsync {
 
-	void sendCode(String string, int points, AsyncCallback<String> callback);
+	void sendCode(String string, Long points, AsyncCallback<String> callback);
 }
